Handle the Event::CONNECTING from PortHanlder
There was a bug, when after socket server disconnection the port closing event was not forwarded to GUI and was
therefore still avilable on activePort comboBox (where actually the data cannot be sent to it).
On the other hand, when the socket client restores the connection back, the port was insterted twice on activePort comboBox.
When PortHandler is switching to Event::CONNECTING state, GUI shall remove it's name from comboBox.