Fix of buffer overflow when reading data in delimiter mode in SocketDrivers
When there was pending data in recv_buffer without delimiter, than reading of size SOCKET_MAX_PAYLOAD_LENGTH caused the recv_buffer overflow.
Currently, when there is data in recv_buffer, the reading size is restricted to fit the buffer size.
When the buffer is full and there is still no delimiter found, entire buffer is abandoned.
Worth to consider, if in case of missing delimiter the buffer should be sent to listeners anyway.