Change TerminalView from QListWidget to QTextEdit
This change is a step backward. QListWidget was an performance overkill in case of high traffic on ports.
Coloring and item-based model was cpu-consuming. Change to QTextEdit significantly increased the performance,
but the disadvantage is missing text/background coloring. This feature can be added when custom model for terminal view will be created.