TestFramework have it's own logger module
Logging module is not independent from the application SW.
Debugs from Drivers and other places have to be removed since these components are reused in TestFramework.
After some time such a debugs can be restored, but the TestLogger have to translate macros from Logger to TestLogger.
RPC was moved to testing directory because is not used in application SW anymore.