Fix SETTINGS persistent storage
When the settings was removed from system, there was a crash when reading from persistence.
Currently, the setting type is stored in persistence so basing on this we are able to read all persistent data related to this setting
like name, type and value and just skip this in case that settings do not exist anymore.