Fix data race when reporting throughput calculation result.
In PortHandler throughput cannot be updated on port close event, because
this method is called from GUI thread. There is possibility, that in the
meantime when port is getting closed, the throughput timer expires and
throughput update is queued in GUI queue (since GUI thread is currently
closing the port). At the end of closing port there was throughput clear
event, which had priority over the queue (the same, GUI thread) and was
executed immediately. Then update queued from timer event was executed,
that is why the throughput was set to non empty value.