UserButtons allows to send data to specified port ID.
By using the '@x>' prefix user can choose, to which port following
command will be sent. For example, following command will be send to
port with ID=0:

@0>some_command

Acceptable ID are from 0 to 4.
In case of typo's in this command, it is treated like a normal string
and send to currently selected port in GUI (default behavior for every
command).