Add COMMAND port type
This port type can be used to gather output from external processes (like 'adb logcat').
This is achieved with use of ProcessDriver which sets up pipes between parent and child process to have access to output of the child.
Child process with user-defined command is started with 'system()' linux call.
When child process is terminated, port behaves as any other type of port and keep restarting the preocess after defined timeout (increased in this commit from 500ms to 5sec).